space/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
timm
tensorflow
Pillow
numpy
huggingface_hub
albumentations
